+++ lib/presentation/cycle/ajout_observation_journee/widget/observation_form.dart
@@ -255,7 +255,7 @@ class _ObservationFormState extends ConsumerState<ObservationForm> {
                 child: TextFormField(
                   keyboardType: TextInputType.number,
                   autocorrect: false,
-                  onChanged: (value) => notifierForm.temperatureBasaleChanged(int.parse(value)),
+                  onChanged: (value) => notifierForm.temperatureBasaleChanged(double.parse(value)),
                   controller: _controllerTempBasale,
                 ),
               ),
